pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "auto-tune-vllm"
version = "0.1.0"
description = "Distributed hyperparameter optimization framework for vLLM serving"
readme = "README.md"
license = {file = "LICENSE"}
authors = [
    {name = "Your Name", email = "your.email@example.com"}
]
keywords = ["vllm", "optimization", "llm", "serving", "benchmarking", "ray"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: Apache Software License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Scientific/Engineering :: Artificial Intelligence",
]
requires-python = ">=3.10"
dependencies = [
    "optuna>=3.0.0",
    "optuna-integration[botorch]>=4.0.0",
    "ray[default]>=2.0.0",
    "vllm>=0.2.0",
    "guidellm>=0.1.0",
    "pyyaml>=6.0",
    "pydantic>=2.0.0",
    "typer>=0.9.0",
    "rich>=13.0.0",
    "requests>=2.28.0",
]

[project.optional-dependencies]
postgresql = [
    "psycopg2-binary>=2.9.0",
]
dev = [
    "pytest>=7.0.0",
    "black>=23.0.0",
    "isort>=5.0.0",
    "mypy>=1.0.0",
    "pre-commit>=3.0.0",
]
viz = [
    "plotly>=5.0.0",
    "pandas>=2.0.0",
]

[project.urls]
Homepage = "https://github.com/your-org/auto-tune-vllm"
Documentation = "https://auto-tune-vllm.readthedocs.io"
Repository = "https://github.com/your-org/auto-tune-vllm"
Issues = "https://github.com/your-org/auto-tune-vllm/issues"

[project.scripts]
auto-tune-vllm = "auto_tune_vllm.cli:main"

[tool.setuptools.packages.find]
where = ["."]
include = ["auto_tune_vllm*"]

[tool.black]
line-length = 88
target-version = ['py310']

[tool.isort]
profile = "black"
line_length = 88

[tool.mypy]
python_version = "3.10"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true